move prisma export to separate file, move import paths
