+++ src/server/src/services/getAllAbilityMetadata.ts
@@ -1,5 +1,5 @@
 import { type Ability } from '@prisma/client';
-import { prisma } from '../index';
+import { prisma } from 'src/db';
 
 export async function getAllAbilityMetadata(): Promise<Ability[]> {
   const results: Ability[] = await prisma.ability.findMany({
