+++ src/server/src/services/getItemMetadata.ts
@@ -1,5 +1,5 @@
 import { type Item } from '@prisma/client';
-import { prisma } from '../index';
+import { prisma } from 'src/db';
 
 export async function getItemMetadata(itemHrid: string): Promise<Item[]> {
   const results: Item[] = await prisma.item.findMany({
